refactor: split out resizable drawer
