+++ apps/jscad-web/examples/hulls.example.js
@@ -13,9 +13,14 @@ const { circle, cuboid, rectangle, sphere } = jscad.primitives
 const { translate } = jscad.transforms
 const { hull, hullChain } = jscad.hulls
 
-const getParameterDefinitions = () => [
-  { name: 'doHull', type: 'radio', caption: 'Show:', values: ['shapes', 'hull', 'chain'], captions: ['Original Shapes', 'Hull', 'Hull Chain'], initial: 'shapes' }
-]
+const getParameterDefinitions = () => [{
+    name: 'doHull',
+    type: 'radio',
+    caption: 'Hull type:',
+    values: ['shapes', 'hull', 'chain'],
+    captions: ['Original Shapes', 'Hull', 'Hull Chain'],
+    initial: 'shapes'
+}]
 
 const main = (params) => {
   const shapes2d = [
