+++ apps/jscad-web/static/index.html
@@ -75,6 +75,7 @@
         <ul>
           <li><a href="/docs/tutorial-01_gettingStarted.html" target="_blank">Getting Started</a></li>
           <li><a href="/docs/" target="_blank">API Reference</a></li>
+          <li><a href="https://github.com/jscad/OpenJSCAD.org" target="_blank">GitHub</a></li>
         </ul>
       </div>
